backend/requirements.txt
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
audioread==3.1.0
babel==2.18.0
backoff==2.2.1
bcrypt==5.0.0
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.0
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
csvw==3.7.0
cymem==2.0.13
decorator==5.2.1
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
fastapi==0.129.0
filelock==3.21.2
flatbuffers==25.12.19
fsspec==2026.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain-core==1.2.12
langchain-text-splitters==1.1.0
langsmith==0.7.1
language-tags==1.2.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
miniaudio==1.61
misaki==0.9.4
mlx==0.30.6
mlx-audio==0.3.1
mlx-lm==0.30.5
mlx-metal==0.30.6
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.15
num2words==0.5.14
numba==0.63.1
numpy==2.3.5
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
phonemizer-fork==3.3.2
platformdirs==4.8.0
pooch==1.9.0
posthog==5.4.0
preshed==3.0.12
protobuf==6.33.5
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyloudnorm==0.2.0
PyMuPDF==1.27.1
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
segments==2.3.0
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
sqlmodel==0.0.33
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.0.0rc3
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
